# openapi.yaml
openapi: 3.0.3
info:
  title: cloudflare-project-2026 API
  description: >
    Sample API surface for the cloudflare-project-2026 template. Replace the
    paths and schemas below with your project's actual contract. This file is
    the canonical source for the API contract — src/routes/ handlers must
    match the paths defined here (see .claude/rules/openapi-sync.md).
  version: 0.1.0

servers:
  - url: http://localhost:8787
    description: Local development (wrangler dev)
  - url: https://cloudflare-project-2026-staging.workers.dev
    description: Staging
  - url: https://cloudflare-project-2026.workers.dev
    description: Production

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      tags: [System]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /api/users:
    get:
      operationId: listUsers
      summary: List users
      tags: [Users]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: createUser
      summary: Create a user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string